pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "joblib>=1.4.2",
    "kafka-python>=2.1.5",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "tensorflow>=2.14.0",
]
